backend/pyproject.toml
[project]
name = "lumipath-backend"
version = "0.1.0"
description = "LumiPath AI-driven personal growth OS — Backend"
requires-python = ">=3.12"
dependencies = [
    # Web framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "python-multipart>=0.0.9",

    # Pydantic & Settings
    "pydantic[email]>=2.7.0",
    "pydantic-settings>=2.3.0",

    # Database (async + sync)
    "sqlalchemy[asyncio]>=2.0.30",
    "asyncpg>=0.29.0",
    "psycopg2-binary>=2.9.9",
    "alembic>=1.13.0",
    "pgvector>=0.3.0",

    # Redis
    "redis[hiredis]>=5.0.0",

    # Neo4j
    "neo4j>=5.20.0",

    # Security & Auth
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "authlib>=1.3.0",
    "httpx>=0.27.0",

    # LLM & Agent
    "langgraph>=0.3.0",
    "langchain-core>=0.3.0",
    "litellm>=1.40.0",

    # Markdown & Vault
    "python-frontmatter>=1.1.0",
    "markdown-it-py>=3.0.0",
    "watchfiles>=0.21.0",
    "gitpython>=3.1.0",
    "jinja2>=3.1.0",

    # Async tasks
    "celery[amqp]>=5.3.0",
    "kombu>=5.3.0",

    # MCP
    "mcp>=1.0.0",

    # Utilities
    "orjson>=3.10.0",
    "structlog>=24.1.0",
    "python-slugify>=8.0.0",
    "click>=8.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "httpx>=0.27.0",
    "factory-boy>=3.3.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
]

[project.scripts]
lumipath = "app.cli.main:cli"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]
ignore = ["B008"]  # FastAPI Depends() in function args

[tool.mypy]
python_version = "3.12"
strict = false
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
